SCS: allow to enter strings

+++ apps/sequencers/midibox_seq_v4/core/seq_lcd.h
@@ -53,7 +53,7 @@ extern s32 SEQ_LCD_InitSpecialChars(seq_lcd_charset_t charset);
 extern s32 SEQ_LCD_PrintString(char *str);
 extern s32 SEQ_LCD_PrintFormattedString(char *format, ...);
 
-extern s32 SEQ_LCD_PrintSpaces(u8 num);
+extern s32 SEQ_LCD_PrintSpaces(int num);
 extern s32 SEQ_LCD_PrintStringPadded(char *str, u32 width);
 extern s32 SEQ_LCD_PrintVBar(u8 value);
 extern s32 SEQ_LCD_PrintHBar(u8 value);
